converted double to float

+++ GameParticles/Matrix.cpp
@@ -141,7 +141,7 @@ void Matrix::set( MatrixRow row, Vect4D *t )
 	}
 }
 
-double &Matrix::operator[](Index e)
+float &Matrix::operator[](Index e)
 {
 	// get the individual elements
 	switch(e)
@@ -265,11 +265,11 @@ Matrix Matrix::operator*( Matrix& rhs)
 	return tmp;
 }
 
-Matrix& Matrix::operator/=(double rhs)
+Matrix& Matrix::operator/=(float rhs)
 { 
 	// divide each element by a value
 	// using inverse multiply trick, faster that individual divides
-	double inv_rhs = 1.0/rhs;
+	float inv_rhs = 1.0f/rhs;
 
 	m0 *= inv_rhs;
 	m1 *= inv_rhs;
@@ -291,19 +291,19 @@ Matrix& Matrix::operator/=(double rhs)
 	return *this;
 }
 
-double Matrix::Determinant() 
+float Matrix::Determinant() 
 {
-	double ta = (m10 * m15) - (m11 * m14);
+	float ta = (m10 * m15) - (m11 * m14);
 	// tb = (kq - mo)
-	double tb = (m9 * m15) - (m11 * m13);
+	float tb = (m9 * m15) - (m11 * m13);
 	// tc = (kp - lo)
-	double tc = (m9 * m14) - (m10 * m13);
+	float tc = (m9 * m14) - (m10 * m13);
 	// td = (jq - mn)
-	double td = (m8 * m15) - (m11 * m12);
+	float td = (m8 * m15) - (m11 * m12);
 	// te = (jo - kn)
-	double te = (m8 * m13) - (m9 *  m12);
+	float te = (m8 * m13) - (m9 *  m12);
 	// tf = (jp - ln)
-	double tf = (m8 * m14) - (m10 * m12);
+	float tf = (m8 * m14) - (m10 * m12);
 	
 	return ((m0 * ((m5 * ta) - (m6 * tb) + (m7 * tc)))
 			- (m1 * ((m4 * ta) - (m6 * td) + (m7 * tf)))
@@ -317,9 +317,9 @@ Matrix Matrix::GetAdjugate()
 	Matrix tmp;
 	
 	// load		ABC		(3)		ABC--
-	double t1 = (m10*m15) - (m11*m14);
-	double t2 = (m9*m15) - (m11*m13);
-	double t3 = (m9*m14) - (m10*m13);
+	float t1 = (m10*m15) - (m11*m14);
+	float t2 = (m9*m15) - (m11*m13);
+	float t3 = (m9*m14) - (m10*m13);
 	
 	// a = f(ta) - g(tb) + h(tc)
 	tmp.m0 = (m5*t1) - (m6*t2) + (m7*t3);
@@ -327,8 +327,8 @@ Matrix Matrix::GetAdjugate()
 	tmp.m1 = -((m1*t1) - (m2*t2) + (m3*t3));
 	
 	// load		JK		(5)		ABCJK
-	double t4 = (m8*m15) - (m11*m12);
-	double t5 = (m8*m14) - (m10*m12);
+	float t4 = (m8*m15) - (m11*m12);
+	float t5 = (m8*m14) - (m10*m12);
 	// e = - ( e(ta) - g(tj) + h(tk))
 	tmp.m4 = -( (m4*t1) - (m6*t4) + (m7*t5));
 	// f = a(ta) - c(tj) + d(tk)
@@ -411,7 +411,7 @@ Matrix Matrix::GetAdjugate()
 void Matrix::Inverse( Matrix &out ) 
 {
 	Matrix tmp;
-	double det = Determinant();
+	float det = Determinant();
 	if(fabs(det) < 0.0001)
 	{
 		// do nothing, Matrix is not invertable
@@ -454,7 +454,7 @@ void Matrix::setScaleMatrix(Vect4D *scale)
 	this->m15 = 1.0;
 }
 
-void Matrix::setRotZMatrix(double az)
+void Matrix::setRotZMatrix(float az)
 {
 	//	{	cos		-sin	0		0	}
 	//	{	sin		cos		0		0	}
