+++ GameParticles/Vect4D.h
@@ -18,23 +18,23 @@ public:
 	friend class Matrix;
 
 	Vect4D();
-	Vect4D(double tx, double ty, double tz, double tw = 1.0f);
+	Vect4D(float tx, float ty, float tz, float tw = 1.0f);
 	~Vect4D();
 
 	void Cross(Vect4D &vin, Vect4D &vout);
-	double &operator[](Vect e);
+	float &operator[](Vect e);
 	void norm(Vect4D &out);
-	void set(double tx, double ty, double tz, double tw = 1.0f);
+	void set(float tx, float ty, float tz, float tw = 1.0f);
 
-	Vect4D operator * (double scale);
+	Vect4D operator * (float scale);
 	Vect4D operator - (Vect4D t);
 	Vect4D operator + (Vect4D t);
 
 private:
-	double x;
-	double y;
-	double z;
-	double w;
+	float x;
+	float y;
+	float z;
+	float w;
 };
 
 #endif
